ノート作成画面のstate管理煩雑なためリファクタリング
